refactor(folio): centralize document enum values
